Custom Present animation을 위한 PaginatingViewControllerPresentAnimator 객체 추가, ArtworkViewController present 및 3D touch를 위한 delegate를 PaginatingCollectionViewController에 추가
