+++ BeBrav/ViewControllers/ArtworkViewController/ArtworkViewController.swift
@@ -298,96 +298,3 @@ extension ArtworkViewController: UIScrollViewDelegate {
         dismiss(animated: true, completion: nil)
     }
 }
-
-
-
-/*
- // 다음 코드들을 메인화면 ViewController에 추가해야함
- 
- override func viewDidLoad() {
- super.viewDidLoad()
- if UIApplication.shared.keyWindow?.traitCollection.forceTouchCapability == .available
- {
- registerForPreviewing(with: self, sourceView: collectionView)
- 
- }
- }
- 
- // MARK:- Return ArtworkViewController
- private func artworkViewController(index: IndexPath) -> ArtworkViewController {
- guard let cell = collectionView.cellForItem(at: index) as? PaginatingCell else {
- return .init()
- }
- 
- let viewController = ArtworkViewController()
- viewController.transitioningDelegate = self
- viewController.mainNavigationController = navigationController
- viewController.artwork = artworkBucket[index.item]
- viewController.artworkImage = cell.artworkImageView.image
- 
- return viewController
- }
- 
- override func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
- let viewController = artworkViewController(index: indexPath)
- viewController.isAnimating = true
- 
- present(viewController, animated: true) {
- viewController.isAnimating = false
- }
- }
- extension PaginatingCollectionViewController: UIViewControllerPreviewingDelegate {
- func previewingContext(_ previewingContext: UIViewControllerPreviewing,
- viewControllerForLocation location: CGPoint)
- -> UIViewController?
- {
- guard let index = collectionView.indexPathForItem(at: location) else {
- return .init()
- }
- let viewController = artworkViewController(index: index)
- viewController.isPeeked = true
- 
- return viewController
- }
- 
- func previewingContext(_ previewingContext: UIViewControllerPreviewing,
- commit viewControllerToCommit: UIViewController)
- {
- guard let viewController = viewControllerToCommit as? ArtworkViewController else {
- return
- }
- viewController.isPeeked = false
- 
- present(viewController, animated: false, completion: nil)
- }
- }
- 
- extension PaginatingCollectionViewController: UIViewControllerTransitioningDelegate {
- func animationController(forPresented presented: UIViewController,
- presenting: UIViewController,
- source: UIViewController)
- -> UIViewControllerAnimatedTransitioning?
- {
- guard let collectionView = collectionView,
- let index = collectionView.indexPathsForSelectedItems?.first,
- let cell = collectionView.cellForItem(at: index)
- else
- {
- return nil
- }
- 
- let transition = PaginatingViewControllerPresentAnimator()
- 
- transition.viewFrame = view.frame
- transition.originFrame = collectionView.convert(cell.frame, to: nil)
- 
- return transition
- }
- 
- func animationController(forDismissed dismissed: UIViewController)
- -> UIViewControllerAnimatedTransitioning?
- {
- return nil
- }
- }
- */
